camel_agent/requirements.txt
camel-ai[all]
python-dotenv>=1.0.0
Pillow>=10.0.0
volcengine-python-sdk[ark]>=1.0.0
flask==3.0.2
flask-cors==4.0.0